+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 ## Unreleased
+
+- Support reply on threads
+
 ## 0.1.1 - 2025-04-02
 
 - Add ./bin/run-ruboty script to run ruboty locally with this gem