+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Unreleased
 
 - Support reply on threads
+- Ignore message events without text field
 
 ## 0.1.1 - 2025-04-02